% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{absmapsdata_file_list}
\alias{absmapsdata_file_list}
\title{Package file list}
\format{
An \code{tibble} object containing the names of files available in the package.
}
\usage{
absmapsdata_file_list
}
\description{
Package file list
}
\keyword{datasets}
