% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{bitre_work_zones2016}
\alias{bitre_work_zones2016}
\title{BITRE Working Zones, 2016}
\format{
An \code{sf} object with 7 variables:
\describe{
\item{\code{wz_code}}{Working zone code}
\item{\code{wz_code16}}{Working zone code 2016}
\item{\code{wz_code16}}{Working zone code 2016}
\item{\code{wz_name16}}{Working zone name 2016}
\item{\code{areasqkm16}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
bitre_work_zones2016
}
\description{
Geospatial data provided by BITRE/The Department of Infrastructure,
Transport, Regional Development, Communications and the Arts
}
\keyword{datasets}
