% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{ced2018}
\alias{ced2018}
\title{Commonwelath Electoral Districts, 2018}
\format{An \code{sf} object with six variables:
#' \describe{
\item{\code{ced_code_2018}}{The CED code}
\item{\code{ced_name_2018}}{The CED name}
\item{\code{albers_sqkm_2018}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}}
\usage{
ced2018
}
\description{
Geospatial data provided by the ABS for Commonwealth Electoral Districts (CEDs) in 2018.
}
\keyword{datasets}
