% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{ced2021}
\alias{ced2021}
\title{Commonwelath Electoral Districts, 2021}
\format{
An \code{sf} object with six variables:
\describe{
\item{\code{ced_code_2021}}{The CED code}
\item{\code{ced_name_2021}}{The CED name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
ced2021
}
\description{
Geospatial data provided by the ABS for Commonwealth Electoral Districts (CEDs) in 2021.
}
\keyword{datasets}
