% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{dz2016}
\alias{dz2016}
\title{Destination Zones, 2016}
\format{
An \code{sf} object with nine variables:
\describe{
\item{\code{dz_code_2016}}{The destination zone code as a character string.}
\item{\code{sa2_code_2016}}{The full 9 digit SA2 code numeric}
\item{\code{sa2_shortcode_2016}}{The 5 digit SA2 code numeric}
\item{\code{sa2_name_2016}}{The SA2 name character}
\item{\code{state_name_2016}}{The full state name}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
dz2016
}
\description{
Geospatial data provided by the ABS for destination zones in 2016.
}
\keyword{datasets}
