% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{dz2021}
\alias{dz2021}
\title{Destination Zones, 2021}
\format{
An \code{sf} object with nine variables:
\describe{
\item{\code{dz_code_2021}}{The destination zone code as a character string.}
\item{\code{sa2_code_2021}}{The 9 digit SA2 code numeric}
\item{\code{sa2_name_2021}}{The SA2 name character}
\item{\code{state_name_2021}}{The full state name}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
dz2021
}
\description{
Geospatial data provided by the ABS for destination zones in 2021.
}
\keyword{datasets}
