% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{employment_regions2015}
\alias{employment_regions2015}
\title{Employment Regions, 2015-2020}
\format{
An \code{sf} object with three variables:
\describe{
\item{\code{employment_region_code_2015}}{The Employment Regions code}
\item{\code{employment_region_name_2015}}{The Employment Regions name}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
employment_regions2015
}
\description{
Geospatial data provided by the Department of Jobs and Small Business for Employment Regions in 2015-2020.
}
\keyword{datasets}
