% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{gcc2011}
\alias{gcc2011}
\title{Greater Capital Cities, 2011}
\format{An \code{sf} object with 10 variables:
#' \describe{
\item{\code{gcc_code_2011}}{The alphanumeric Greater Capital City (GCC) code numeric}
\item{\code{gcc_name_2011}}{The GCC name}
\item{\code{state_name_2011}}{The full state name}
\item{\code{albers_sqkm_2011}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}}
\usage{
gcc2011
}
\description{
Geospatial data provided by the ABS for Greater Capital Cities (GCC) in 2011.
}
\keyword{datasets}
