% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{gcc2021}
\alias{gcc2021}
\title{Greater Capital Cities, 2021}
\format{
An \code{sf} object with 12 variables:
\describe{
\item{\code{gcc_code_2021}}{The alphanumeric Greater Capital City (GCC) code character}
\item{\code{gcc_name_2021}}{The GCC name}
\item{\code{state_name_2021}}{The full state name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
gcc2021
}
\description{
Geospatial data provided by the ABS for Greater Capital Cities (GCC) in 2021. Note that the Geocentric Datum of Australia 2020 (GDA2020) is used.
See [ABS explanation](https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files).
}
\keyword{datasets}
