% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{iare2021}
\alias{iare2021}
\title{Indigenous Areas (IAREs), 2021}
\format{
An \code{sf} object with nine variables:
\describe{
\item{\code{iar_code_2021}}{The Indigenous Areas code}
\item{\code{iar_name_2021}}{The Indigenous Areas name}
\item{\code{ire_code_2021}}{The Indigenous Regions code}
\item{\code{ire_name_2021}}{The Indigenous Regions name}
\item{\code{state_code_2021}}{The state code}
\item{\code{state_name_2021}}{The full state name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\source{
<https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files>
}
\usage{
iare2021
}
\description{
Geospatial data provided by the ABS for IAREs in 2021, where 'Indigenous Areas (IAREs) are medium sized geographical areas designed to facilitate the release of more detailed statistics for Aboriginal and Torres Strait Islander people.'
}
\keyword{datasets}
