% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{iloc2021}
\alias{iloc2021}
\title{Indigenous Locations (ILOCs), 2021}
\format{
An \code{sf} object with 11 variables:
\describe{
\item{\code{ilo_code_2021}}{The Indigenous Locations code}
\item{\code{ilo_name_2021}}{The Indigenous Locations name}
\item{\code{iar_code_2021}}{The Indigenous Areas code}
\item{\code{iar_name_2021}}{The Indigenous Areas name}
\item{\code{ire_code_2021}}{The Indigenous Regions code}
\item{\code{ire_name_2021}}{The Indigenous Regions name}
\item{\code{state_code_2021}}{The state code}
\item{\code{state_name_2021}}{The full state name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\source{
<https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files>
}
\usage{
iloc2021
}
\description{
Geospatial data provided by the ABS for ILOCs in 2021, where 'Indigenous Locations (ILOCs) represent small Aboriginal and Torres Strait Islander communities (urban and rural) with a minimum population of about 90 people.'
}
\keyword{datasets}
