% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{ireg2021}
\alias{ireg2021}
\title{Indigenous Regions (IREGs), 2021}
\format{
An \code{sf} object with seven variables:
\describe{
\item{\code{ire_code_2021}}{The Indigenous Regions code}
\item{\code{ire_name_2021}}{The Indigenous Regions name}
\item{\code{state_code_2021}}{The state code}
\item{\code{state_name_2021}}{The full state name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\source{
<https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files>
}
\usage{
ireg2021
}
\description{
Geospatial data provided by the ABS for IREGs in 2021, where 'Indigenous Regions (IREGs) are large geographical areas based on historical boundaries. The larger population of Indigenous Regions enables highly detailed analysis.'
}
\keyword{datasets}
