% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{lga2022}
\alias{lga2022}
\title{Local Government Areas, 2022}
\format{
An \code{sf} object with eight variables:
\describe{
\item{\code{lga_code_2022}}{The LGA code}
\item{\code{lga_name_2022}}{The LGA name}
\item{\code{state_code_2022}}{The state code}
\item{\code{state_name_2022}}{The full state name}
\item{\code{areasqkm_2022}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
lga2022
}
\description{
Geospatial data provided by the ABS for local government areas (LGAs) in 2022.
}
\keyword{datasets}
