% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{nsw_lhd2023}
\alias{nsw_lhd2023}
\title{NSW Local Health Districts, 2023}
\format{
An \code{sf} object with two variables:
\describe{
\item{\code{nsw_lhd_name}}{The NSW Local Health District name}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\source{
<https://www.google.com/maps/d/u/0/viewer?mid=1Dv1JRTGmzlm83tBv7tb8vQcOQXY&hl=en_US&ll=-36.014960831899266%2C146.956844&z=9>
}
\usage{
nsw_lhd2023
}
\description{
Geospatial data of NSW Local Health Districts (see source)
}
\keyword{datasets}
