% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{ra2011}
\alias{ra2011}
\title{Remoteness areas, 2011}
\format{An \code{sf} object with 10 variables:
#' \describe{
\item{\code{ra_code}}{The remoteness area code (by state)}
\item{\code{ra}}{The remoteness area name (by state)}
\item{\code{albers_sqkm_2011}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}}
\usage{
ra2011
}
\description{
Geospatial data provided by the ABS for remoteness areas in 2011.
}
\keyword{datasets}
