% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{rda2016}
\alias{rda2016}
\title{Regional Development Australia 2015-16}
\format{
An \code{sf} object with two variables:
\describe{
\item{\code{rda_code_2016}}{Regional Development Australia code}
\item{\code{rda_name_2016}}{Regional Development Australia name}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\source{
<https://data.gov.au/data/dataset/rda-2015-16>
}
\usage{
rda2016
}
\description{
RDA Regions are created and maintained by the Dept. of Infrastructure and Regional Development to support this initiative.
}
\keyword{datasets}
