% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{sa12016}
\alias{sa12016}
\title{Statistical Area 1, 2016}
\format{An \code{sf} object with 17 variables:
#' \describe{
\item{\code{sa1_code_2016}}{The full 11 digit SA1 code numeric}
\item{\code{sa1_7dig_2016}}{The 7 digit SA1 code numeric}
\item{\code{sa2_code_2016}}{The full 9 digit SA2 code numeric}
\item{\code{sa2_shortcode_2016}}{The 5 digit SA2 code numeric}
\item{\code{sa2_name_2016}}{The SA2 name character}
\item{\code{sa3_code_2016}}{The full 5 digit SA3 code numeric}
\item{\code{sa3_name_2016}}{The SA3 name character}
\item{\code{sa4_code_2016}}{The full 3 digit SA4 code numeric}
\item{\code{sa4_name_2016}}{The SA4 name character}
\item{\code{gcc_code_2016}}{The alphanumeric Greater Capital City (GCC) code numeric}
\item{\code{gcc_name_2016}}{The GCC name}
\item{\code{state_name_2016}}{The full state name}
\item{\code{albers_sqkm_2016}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}}
\usage{
sa12016
}
\description{
Geospatial data provided by the ABS for Statistical Area 1 in 2016.
}
\keyword{datasets}
