% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{sa42011}
\alias{sa42011}
\title{Statistical Area 4, 2011}
\format{An \code{sf} object with 10 variables:
#' \describe{
\item{\code{sa4_code_2011}}{The full 3 digit SA4 code numeric}
\item{\code{sa4_name_2011}}{The SA4 name character}
\item{\code{gcc_code_2011}}{The alphanumeric Greater Capital City (GCC) code numeric}
\item{\code{gcc_name_2011}}{The GCC name}
\item{\code{state_name_2011}}{The full state name}
\item{\code{albers_sqkm_2011}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}}
\usage{
sa42011
}
\description{
Geospatial data provided by the ABS for Statistical Area 4 in 2011.
}
\keyword{datasets}
