% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{sed2018}
\alias{sed2018}
\title{State Electoral Districts, 2018}
\format{An \code{sf} object with six variables:
#' \describe{
\item{\code{sed_code_2018}}{The SED code}
\item{\code{sed_name_2018}}{The SED name}
\item{\code{albers_sqkm_2018}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}}
\usage{
sed2018
}
\description{
Geospatial data provided by the ABS for State Electoral Districts (SEDs) in 2018.
}
\keyword{datasets}
