% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{sed2021}
\alias{sed2021}
\title{State Electoral Districts, 2021}
\format{
An \code{sf} object with six variables:
\describe{
\item{\code{sed_code_2021}}{The SED code}
\item{\code{sed_name_2021}}{The SED name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
sed2021
}
\description{
Geospatial data provided by the ABS for State Electoral Districts (SEDs) in 2021.
}
\keyword{datasets}
