% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{sos2016}
\alias{sos2016}
\title{Section of State (SOS), 2016}
\format{
An \code{sf} object with eight variables:
\describe{
\item{\code{sos_code_2016}}{The SOS code}
\item{\code{sos_name_2016}}{The SOS name}
\item{\code{state_code_2016}}{The state code}
\item{\code{state_name_2016}}{The full state name}
\item{\code{areasqkm_2016}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
sos2016
}
\description{
Geospatial data provided by the ABS for Section of State (SOS) in 2016.
}
\keyword{datasets}
