% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{sosr2016}
\alias{sosr2016}
\title{Section of State Range (SOSR), 2016}
\format{
An \code{sf} object with ten variables:
\describe{
\item{\code{ssr_code_2016}}{The SSR code}
\item{\code{ssr_name_2016}}{The SSR name}
\item{\code{sos_code_2016}}{The SOS code}
\item{\code{sos_name_2016}}{The SOS name}
\item{\code{state_code_2016}}{The state code}
\item{\code{state_name_2016}}{The full state name}
\item{\code{areasqkm_2016}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
sosr2016
}
\description{
Geospatial data provided by the ABS for Section of State Range (SOSR) in 2016.
}
\keyword{datasets}
