% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{state2021}
\alias{state2021}
\title{States, 2021}
\format{
An \code{sf} object with 10 variables:
\describe{
\item{\code{state_name_2021}}{The full state name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
state2021
}
\description{
Geospatial data provided by the ABS for states in 2021. Note that the Geocentric Datum of Australia 2020 (GDA2020) is used.
See [ABS explanation](https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/digital-boundary-files).
}
\keyword{datasets}
