% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{sua2016}
\alias{sua2016}
\title{Significant Urban Area (SUA), 2016}
\format{
An \code{sf} object with six variables:
\describe{
\item{\code{sua_code_2016}}{The SUA code}
\item{\code{sua_name_2016}}{The SUA name}
\item{\code{areasqkm_2016}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
sua2016
}
\description{
Geospatial data provided by the ABS for Significant Urban Areas (SUA) in 2016.
}
\keyword{datasets}
