% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{sua2021}
\alias{sua2021}
\title{Significant Urban Area (SUA), 2021}
\format{
An \code{sf} object with six variables:
\describe{
\item{\code{sua_code_2021}}{The SUA code}
\item{\code{sua_name_2021}}{The SUA name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
sua2021
}
\description{
Geospatial data provided by the ABS for Significant Urban Areas (SUA) in 2021.
}
\keyword{datasets}
