% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{suburb2016}
\alias{suburb2016}
\title{Suburbs, 2016}
\format{
An \code{sf} object with five variables:
\describe{
\item{\code{suburb_2016}}{The suburb (SSC) number as a character string.}
\item{\code{suburb_num_2016}}{The suburb number as a numeric.}
\item{\code{state_code_2016}}{The state code.}
\item{\code{state_name_2016}}{The full state name.}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
suburb2016
}
\description{
Geospatial data provided by the ABS for State Suburbs (SSC) in 2016.
}
\keyword{datasets}
