% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{suburb2021}
\alias{suburb2021}
\title{Suburbs, 2021}
\format{
An \code{sf} object with five variables:
\describe{
\item{\code{suburb_2021}}{The suburb (SAL) number as a character string.}
\item{\code{suburb_num_2021}}{The suburb number as a numeric.}
\item{\code{state_code_2021}}{The state code.}
\item{\code{state_name_2021}}{The full state name.}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
suburb2021
}
\description{
Geospatial data provided by the ABS for Suburbs and Localities (SAL) in 2021.
}
\keyword{datasets}
