% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{tourism2021}
\alias{tourism2021}
\title{Tourism Regions, 2021}
\format{
An \code{sf} object with eight variables:
\describe{
\item{\code{tourism_code_2021}}{The Tourism Regions code}
\item{\code{tourism_name_2021}}{The Tourism Regions name}
\item{\code{state_code_2021}}{The state code}
\item{\code{state_name_2021}}{The full state name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
tourism2021
}
\description{
Geospatial data provided by the ABS for Tourism Regions in 2021.
}
\keyword{datasets}
