% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{ucl2016}
\alias{ucl2016}
\title{Urban Centre and Locality (UCL), 2016}
\format{
An \code{sf} object with twelve variables:
\describe{
\item{\code{ucl_code_2016}}{The UCL code}
\item{\code{ucl_name_2016}}{The UCL name}
\item{\code{ssr_code_2016}}{The SSR code}
\item{\code{ssr_name_2016}}{The SSR name}
\item{\code{sos_code_2016}}{The SOS code}
\item{\code{sos_name_2016}}{The SOS name}
\item{\code{state_code_2016}}{The state code}
\item{\code{state_name_2016}}{The full state name}
\item{\code{areasqkm_2016}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
ucl2016
}
\description{
Geospatial data provided by the ABS for Urban Centre and Locality (UCL) in 2016.
}
\keyword{datasets}
