% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_dat.R
\docType{data}
\name{ucl2021}
\alias{ucl2021}
\title{Urban Centre and Locality (UCL), 2021}
\format{
An \code{sf} object with twelve variables:
\describe{
\item{\code{ucl_code_2021}}{The UCL code}
\item{\code{ucl_name_2021}}{The UCL name}
\item{\code{ssr_code_2021}}{The SSR code}
\item{\code{ssr_name_2021}}{The SSR name}
\item{\code{sos_code_2021}}{The SOS code}
\item{\code{sos_name_2021}}{The SOS name}
\item{\code{state_code_2021}}{The state code}
\item{\code{state_name_2021}}{The full state name}
\item{\code{areasqkm_2021}}{The area in square kilometres}
\item{\code{cent_lat}}{The latitide of the area's centroid}
\item{\code{cent_long}}{The latitide of the area's centroid}
\item{\code{geometry}}{A nested list containing the area's geometry (polygons)}
}
}
\usage{
ucl2021
}
\description{
Geospatial data provided by the ABS for Urban Centre and Locality (UCL) in 2021.
}
\keyword{datasets}
