% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correspondence_absmaps.R
\name{get_correspondence_absmaps}
\alias{get_correspondence_absmaps}
\title{Read in ASGC/ASGS population-weighted correspondence tables from the ABS.}
\usage{
get_correspondence_absmaps(fromArea, fromYear, toArea, toYear)
}
\arguments{
\item{fromArea}{The area you want to correspond FROM (ie the areas your data are currently in). For example: "sa1", "sa2, "sa3", "sa4".}

\item{fromYear}{The year you want to correspond FROM. For example: 2011, 2016.}

\item{toArea}{The area you want to correspond TO (ie the areas you want your data to be in).}

\item{toYear}{The year you want to correspond TO.}
}
\value{
A \code{tibble} object.
}
\description{
Read in ASGC/ASGS population-weighted correspondence tables from the ABS.
}
\examples{
\dontrun{
sa4_corr <- get_correspondence_absmaps("sa4", 2011,
                                       "sa4", 2016)
}
}
